refactor: decompose diagnostics_shared into order_bands, peak_classification, severity
Break up the 495-line grab-bag module into three focused, single-responsibility
modules:

- order_bands.py: order-band math, diagnostic settings, vehicle-order frequencies
- peak_classification.py: peak-frequency classification against known order bands
- severity.py: severity-bucket state tracking with hysteresis and persistence

Changes:
- Delete diagnostics_shared.py
- Migrate all production and test imports to the new modules
- Rename tests/live_diagnostics/ -> tests/diagnostics/ (reflective of actual coverage)
- Update mypy enforcement list in pyproject.toml (3 new modules replace 1)
- Update backend.instructions.md, testing.md, metrics.md docs
- Delete chunk planning files

Validation: ruff clean, mypy clean (133 files), 5015 tests pass, 0 skipped.
