+++ .github/instructions/backend.instructions.md
@@ -12,7 +12,7 @@ Backend (python `apps/server/`)
 	- `apps/server/vibesensor/report/pdf_engine.py`: public PDF renderer entrypoint and page orchestration.
 	- `apps/server/vibesensor/update/`: wheel-based updater package; `manager.py` is the public facade over focused modules for workflow, Wi-Fi, releases, install and rollback, service control, and status.
 - Install: `python -m pip install -e "./apps/server[dev]"` (used by CI).
-- Backend type gate: `make typecheck-backend` runs the enforced mypy slice for app/bootstrap, runtime/routes, core typed-boundary modules (`backend_types.py`, `config.py`, `domain_models.py`, `registry.py`, `settings_store.py`, `api_models.py`, `history_{helpers,runs,reports,exports}.py`, `firmware_cache.py`, `gps_speed.py`, `run_context.py`, `runlog.py`, `esp_flash_manager.py`, `json_types.py`, `payload_types.py`, `report_i18n.py`, `ws_models.py`, `diagnostics_shared.py`, `analysis_persistence.py`, `analysis_settings.py`), and the high-risk `analysis/`, `processing/`, `history_db/`, `metrics_log/`, and `update/` packages.
+- Backend type gate: `make typecheck-backend` runs the enforced mypy slice for app/bootstrap, runtime/routes, core typed-boundary modules (`backend_types.py`, `config.py`, `domain_models.py`, `registry.py`, `settings_store.py`, `api_models.py`, `history_{helpers,runs,reports,exports}.py`, `firmware_cache.py`, `gps_speed.py`, `run_context.py`, `runlog.py`, `esp_flash_manager.py`, `json_types.py`, `payload_types.py`, `report_i18n.py`, `ws_models.py`, `order_bands.py`, `peak_classification.py`, `severity.py`, `analysis_persistence.py`, `analysis_settings.py`), and the high-risk `analysis/`, `processing/`, `history_db/`, `metrics_log/`, and `update/` packages.
 - Prefer explicit payload contracts (`TypedDict`, dataclass, protocol, `JsonValue`/`JsonObject` aliases) over broad `Any` when shaping analysis, report, and persistence data.
 - Treat `Any` as a design smell by default: prefer `object` for untrusted inputs, shared JSON aliases for persisted payloads, `ParamSpec` for callable wrappers, and focused `TypedDict`/protocol contracts for nested state.
 - For live processing / WebSocket payloads, prefer shared contracts in `apps/server/vibesensor/payload_types.py` and `vibesensor_core.vibration_strength` over ad-hoc `dict[str, Any]` bags.
